refactor(plugin): extract color_converter.rs (#1079)
Moves the RhiColorConverterMethodsVTable callbacks (Linux + non-Linux
stubs) + HOST_RHI_COLOR_CONVERTER_METHODS_VTABLE + accessor + both
test modules (gpu_rhi_color_converter_methods_vtable_null_tests +
rhi_color_converter_methods_vtable_tier1_wire_format_tests) out of
host_services/mod.rs into a sibling submodule. Cleans up two
dangling OutputWriter/InputMailboxes test-mod doc comments left
behind by earlier extractions. mod.rs re-exports
HOST_RHI_COLOR_CONVERTER_METHODS_VTABLE and
host_rhi_color_converter_methods_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
